+++ app/human_resources/tests/unit/employee/conftest.py
@@ -16,3 +16,9 @@ def model_kwargs(request, kwargs_employee):
 
     if hasattr(request.cls, 'kwargs_create_item'):
         del request.cls.kwargs_create_item
+
+
+@pytest.fixture( scope = 'class')
+def model_serializer(serializer_employee):
+
+    yield serializer_employee